# Release google-cloud-bigtable 1.0.0